+++ packages/ts-components/src/index.ts
@@ -26,3 +26,8 @@ export {
   AlgoliaSearchProvider,
   useAlgoliaSearch
 } from './helpers/algolia/AlgoliaSearchProvider';
+
+export * from './components/opta/football/fixtures/OptaFootballFixtures';
+export * from './components/opta/football/standings/OptaFootballStandings';
+export * from './components/opta/football/summary/OptaFootballSummary';
+export * from './components/opta/football/match-stats/OptaFootballMatchStats';