+++ packages/ts-components/src/types/externs.d.ts
@@ -69,9 +69,10 @@ declare module '@times-components/styleguide' {
 
   type Fonts = {
     body: string;
+    bodyRegular: string;
     headline: string;
-    supporting: string;
     headlineRegular: string;
+    supporting: string;
   };
 
   export const fonts: Fonts;
@@ -140,4 +141,15 @@ declare module '@times-components/provider-queries' {
 
 declare module '@times-components/provider-test-tools' {
   export const MockedProvider: React.FC<{ mocks: any }>;
+}
+
+declare module '@times-components/date-publication' {
+  type DatePublicationProps = {
+    date: string;
+    publication?: 'SUNDAYTIMES' | 'TIMES';
+    showDay: boolean;
+  };
+
+  export const DatePublication: React.FC<DatePublicationProps>;
+  export default DatePublication;
 }