+++ packages/article-main-comment/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.28.11](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.28.10...@times-components/article-main-comment@2.28.11) (2021-06-17)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.28.10](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.28.9...@times-components/article-main-comment@2.28.10) (2021-06-16)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.28.9](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.28.8...@times-components/article-main-comment@2.28.9) (2021-06-14)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.28.8](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.28.7...@times-components/article-main-comment@2.28.8) (2021-06-10)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.28.7](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.28.6...@times-components/article-main-comment@2.28.7) (2021-06-10)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.28.6](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.28.5...@times-components/article-main-comment@2.28.6) (2021-06-04)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
 ## [2.28.5](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.28.4...@times-components/article-main-comment@2.28.5) (2021-06-04)
 
 **Note:** Version bump only for package @times-components/article-main-comment