+++ packages/message-bar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.63](https://github.com/newsuk/times-components/compare/@times-components/message-bar@0.3.62...@times-components/message-bar@0.3.63) (2021-06-04)
+
+**Note:** Version bump only for package @times-components/message-bar
+
+
+
+
+
 ## [0.3.62](https://github.com/newsuk/times-components/compare/@times-components/message-bar@0.3.61...@times-components/message-bar@0.3.62) (2021-05-27)
 
 **Note:** Version bump only for package @times-components/message-bar