+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.16](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.15...@times-components/pages@2.3.16) (2021-06-17)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.15](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.14...@times-components/pages@2.3.15) (2021-06-16)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.14](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.13...@times-components/pages@2.3.14) (2021-06-14)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.13](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.12...@times-components/pages@2.3.13) (2021-06-10)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.12](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.11...@times-components/pages@2.3.12) (2021-06-10)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.11](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.10...@times-components/pages@2.3.11) (2021-06-04)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
 ## [2.3.10](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.9...@times-components/pages@2.3.10) (2021-06-04)
 
 **Note:** Version bump only for package @times-components/pages